#52 Implement method for getting the child node parents for H2 data source
As H2 does not support PostgreSQL recursive queries, we need to implement functionality for emitting the specified content record parents for this type of data source.

H2 can be used as a master db in case when you want to run an application (doesn't matter containerized or not) as the only component for demonstration purposes.

Records that are being emitted aren't stored in Content cache.

There is a couple of small SQL code optimizations.